refactor: event detail
