--- src/components/EventDetail/FeaturedPerformers.tsx
@@ -1,9 +1,0 @@
-import { Performer } from '@/types/Performer'
-import React from 'react'
-import PerformersSection from '../Home/components/PerformersSection'
-
-const FeaturedPerformers = ({ performers }: { performers: Performer[] }) => {
-  return <PerformersSection performers={performers} />
-}
-
-export default FeaturedPerformers
