+++ src/components/EventDetail/SeatReservation/ConfirmOrderModal.tsx
@@ -9,7 +9,7 @@ import { RadioGroup, RadioGroupItem } from '@/components/ui/radio-group'
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from '@/components/ui/dialog'
 import { Check, CreditCard, Info, X, Loader2 } from 'lucide-react'
 import { useForm } from 'react-hook-form'
-import { SelectedSeat } from './types'
+import { SelectedSeat } from '../types'
 import { useToast } from '@/hooks/use-toast'
 import axios from 'axios'
 import { Event } from '@/payload-types'
