--- src/components/concert-detail/data/seat-maps/categories.ts
@@ -1,32 +1,0 @@
-export const categories = [
-  {
-    id: 'zone1',
-    name: 'Zone 1',
-    color: '#B60208',
-    textColor: '#fff',
-  },
-  {
-    id: 'zone2',
-    name: 'Zone 2',
-    color: '#F99446',
-    textColor: '#fff',
-  },
-  {
-    id: 'zone3',
-    name: 'Zone 3',
-    color: '#EADD1C',
-    textColor: '#fff',
-  },
-  {
-    id: 'zone4',
-    name: 'Zone 4',
-    color: '#1EB0EF',
-    textColor: '#fff',
-  },
-  {
-    id: 'zone5',
-    name: 'Zone 5',
-    color: '#0FAD4F',
-    textColor: '#fff',
-  },
-]
